Extract Configuration class
