+++ lib/shutterstock-ruby.rb
@@ -1,5 +1,6 @@
 require 'json'
 require 'rest_client'
+require 'shutterstock-ruby/configuration'
 require 'shutterstock-ruby/connections'
 require 'shutterstock-ruby/images'
 require 'shutterstock-ruby/videos'
@@ -17,18 +18,5 @@ module ShutterstockRuby
       @images = Images.new(configuration)
     end
 
-    # Main configuration class.
-    class Configuration
-      attr_accessor :access_token
-      attr_accessor :api_client
-      attr_accessor :api_secret
-
-      def initialize(args)
-        @access_token = args[:access_token]
-        @api_client = args[:api_client]
-        @api_secret = args[:api_secret]
-      end
-    end
   end
-
 end
